test(google): split provider test projects
Align the test assemblies with the Google Firestore provider package names and use a pinned official Firebase CLI package in CI.


Copilot-Session: d6afe8bb-ecfb-403d-8566-b0fae481db72
