+++ src/Google/Orleans.Clustering.GoogleFirestore/Orleans.Clustering.GoogleFirestore.csproj
@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <PackageId>Microsoft.Orleans.Clustering.GoogleFirestore</PackageId>
-    <Title>Microsoft Orleans Google Firestore Clustering Provider</Title>
-    <Description>Microsoft Orleans clustering provider backed by Google Firestore</Description>
-    <PackageTags>$(PackageTags) Google Firestore</PackageTags>
+    <Title>Microsoft Orleans Google Cloud Firestore Clustering Provider</Title>
+    <Description>Microsoft Orleans clustering provider backed by Google Cloud Firestore</Description>
+    <PackageTags>$(PackageTags) Google Cloud Firestore</PackageTags>
     <TargetFrameworks>$(DefaultTargetFrameworks)</TargetFrameworks>
     <AssemblyName>Orleans.Clustering.GoogleFirestore</AssemblyName>
     <RootNamespace>Orleans.Clustering.GoogleFirestore</RootNamespace>
@@ -26,8 +26,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>Orleans.Tests.Google</_Parameter1>
-    </AssemblyAttribute>
+    <InternalsVisibleTo Include="Orleans.Clustering.GoogleFirestore.Tests" />
   </ItemGroup>
 </Project>
