+++ src/Google/Orleans.GrainDirectory.GoogleFirestore/Orleans.GrainDirectory.GoogleFirestore.csproj
@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <PackageId>Microsoft.Orleans.GrainDirectory.GoogleFirestore</PackageId>
-    <Title>Microsoft Orleans Google Firestore Grain Directory Provider</Title>
-    <Description>Microsoft Orleans GrainD irectory provider backed by Google Firestore</Description>
-    <PackageTags>$(PackageTags) Google Firestore</PackageTags>
+    <Title>Microsoft Orleans Google Cloud Firestore Grain Directory Provider</Title>
+    <Description>Microsoft Orleans grain directory provider backed by Google Cloud Firestore</Description>
+    <PackageTags>$(PackageTags) Google Cloud Firestore</PackageTags>
     <TargetFrameworks>$(DefaultTargetFrameworks)</TargetFrameworks>
     <AssemblyName>Orleans.GrainDirectory.GoogleFirestore</AssemblyName>
     <RootNamespace>Orleans.GrainDirectory.GoogleFirestore</RootNamespace>
@@ -26,8 +26,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-      <_Parameter1>Orleans.Tests.Google</_Parameter1>
-    </AssemblyAttribute>
+    <InternalsVisibleTo Include="Orleans.GrainDirectory.GoogleFirestore.Tests" />
   </ItemGroup>
 </Project>
