+++ src/Orleans.Core/Orleans.Core.csproj
@@ -72,7 +72,7 @@
     <InternalsVisibleTo Include="Orleans.DefaultCluster.Tests" />
     <InternalsVisibleTo Include="Orleans.DurableJobs" />
     <InternalsVisibleTo Include="Orleans.DurableJobs.Tests" />
-    <InternalsVisibleTo Include="Orleans.Tests.Google" />
+    <InternalsVisibleTo Include="Orleans.Persistence.GoogleFirestore.Tests" />
     <InternalsVisibleTo Include="Orleans.GrainDirectory.Tests" />
     <InternalsVisibleTo Include="Orleans.Placement.Tests" />
     <InternalsVisibleTo Include="Orleans.Redis.Tests" />
