+++ src/Orleans.TestingHost/Orleans.TestingHost.csproj
@@ -25,7 +25,7 @@
     <InternalsVisibleTo Include="Orleans.GrainDirectory.Tests" />
     <InternalsVisibleTo Include="Orleans.Placement.Tests" />
     <InternalsVisibleTo Include="Orleans.Runtime.Internal.Tests" />
-    <InternalsVisibleTo Include="Orleans.Tests.Google" />
+    <InternalsVisibleTo Include="Orleans.Persistence.GoogleFirestore.Tests" />
     <InternalsVisibleTo Include="Orleans.Streaming.Tests" />
   </ItemGroup>
 </Project>
