--- test/Extensions/Google.Tests/Usings.cs
@@ -1,1 +1,0 @@
-global using Xunit;
