+++ test/Extensions/GoogleFirestore/Usings.cs
@@ -1,0 +1,2 @@
+global using Orleans.Tests.GoogleFirestore;
+global using Xunit;
