+++ test/Extensions/Orleans.GrainDirectory.GoogleFirestore.Tests/Orleans.GrainDirectory.GoogleFirestore.Tests.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>$(TestTargetFrameworks)</TargetFrameworks>
+    <OrleansBuildTimeCodeGen>true</OrleansBuildTimeCodeGen>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="$(SourceRoot)src\Google\Orleans.GrainDirectory.GoogleFirestore\Orleans.GrainDirectory.GoogleFirestore.csproj" />
+    <ProjectReference Include="$(SourceRoot)test\Orleans.Runtime.Tests\Orleans.Runtime.Tests.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="$(SourceRoot)test\Extensions\GoogleFirestore\GoogleEmulatorHost.cs" Link="GoogleEmulatorHost.cs" />
+    <Compile Include="$(SourceRoot)test\Extensions\GoogleFirestore\Usings.cs" Link="Usings.cs" />
+  </ItemGroup>
+</Project>
