+++ test/TestInfrastructure/TestExtensions/TestExtensions.csproj
@@ -30,7 +30,7 @@
     <InternalsVisibleTo Include="Orleans.Cosmos.Tests" />
     <InternalsVisibleTo Include="Orleans.Redis.Tests" />
     <InternalsVisibleTo Include="Orleans.Runtime.Internal.Tests" />
-    <InternalsVisibleTo Include="Orleans.Tests.Google" />
+    <InternalsVisibleTo Include="Orleans.Persistence.GoogleFirestore.Tests" />
     <InternalsVisibleTo Include="Orleans.Streaming.NATS.Tests" />
   </ItemGroup>
 </Project>
